+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 - Fix `slumber generate curl` output for multipart forms with file fields
+- `slumber import insomnia` now imports some dynamic expressions. Values chained from other responses now properly import as `response()`/`response_header()` calls [#164](https://github.com/LucasPickering/slumber/issues/164)
 
 ## [4.0.1] - 2025-09-14